Stop checking Trac to skip tests against open tickets.
In general, skipped tests should live only in patches, which are committed at
the same time that the corresponding bug is fixed. In cases where it's
necessary to skip a test, use `markTestSkipped()` to declare this fact
explicitly.

We continue to check Trac when using `WP_UnitTestCase` to run non-core tests.

See #30284.
Built from https://develop.svn.wordpress.org/trunk@30526


git-svn-id: http://core.svn.wordpress.org/trunk@30515 1a063a9b-81f0-0310-95a4-ce76da25c4cd