Add test for `get_term_by( 'slug' )` behavior with accented chars.
When passing a value of 'slug' to `wp_insert_term()` that contains non-ASCII
characters, WordPress converts accented characters to non-accented versions.
(See `sanitize_title()` and `remove_accents()`.) The same conversion happens
when fetching when fetching a term using a slug. In this way, it's possible to
create a term and fetch it using the same accented string, even though the slug
is actually stored with non-accented characters.

See #16282.
Built from https://develop.svn.wordpress.org/trunk@34628


git-svn-id: http://core.svn.wordpress.org/trunk@34592 1a063a9b-81f0-0310-95a4-ce76da25c4cd