Customize: Fix sub-nav menu items from appearing to bump to root level in preview upon saving.
This issue would appear when the parent nav menu item was also newly-inserted. The issue was only apparent in the prevew, as the actual data saved was correct. The `menu_item_parent` properties of the saved sub-nav menu items needed to get updated to refer to the newly-inserted parent `nav_menu_item` post IDs, as opposed to retaining the placeholder IDs.

Fixes #34628.

Built from https://develop.svn.wordpress.org/trunk@35581


git-svn-id: http://core.svn.wordpress.org/trunk@35545 1a063a9b-81f0-0310-95a4-ce76da25c4cd