Emoji: Add support for the Rainbow and Pirate flag emoji.
Twemoji recently added images for the Rainbow and Pirate flags, and the latest iOS 10 beta added support for the Rainbow flag. Never let it be said that WordPress is at anything less than the cutting edge of emoji research and development.

Merge of [38179] to the 4.6 branch.

Props ocean90 for the code review.
Fixes #37543.


Built from https://develop.svn.wordpress.org/branches/4.6@38180


git-svn-id: http://core.svn.wordpress.org/branches/4.6@38121 1a063a9b-81f0-0310-95a4-ce76da25c4cd