External Libraries: Update minified version of jquery.masonry.js.
WordPress maintains the minified version of jquery.masonry.js since there is no official build, however it has been excluded from grunt's minification process. This adds a minification task to grunt, adds it to the precommit hook for JS, minifies the file, and bumps the version on jquery.masonry.min.js. The change to the non minified version was introduced in [38261].

Merge of [38276] and [38281] to the 4.6 branch.

Fixes #37720.  See #37666.
Built from https://develop.svn.wordpress.org/branches/4.6@38339


git-svn-id: http://core.svn.wordpress.org/branches/4.6@38280 1a063a9b-81f0-0310-95a4-ce76da25c4cd